# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SentinelAML API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------
  # Customers
  # ---------------------------------------------------------------------
...
  /analytics/shortest-path:
    get:
      operationId: getShortestPath
      tags: [analytics]
      summary: Shortest path between two accounts
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string }
        - name: to
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Shortest path result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortestPathResult"
  /analytics/most-connected:
    get:
      operationId: getMostConnected
      tags: [analytics]
      summary: Top N accounts by degree
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
      responses:
        "200":
          description: Most connected accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CentralityEntry"

  # ---------------------------------------------------------------------
  # Risk
  # ---------------------------------------------------------------------
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts
      parameters:
        - name: severity
          in: query
          schema:
            $ref: "#/components/schemas/RiskLevel"
        - name: resolved
          in: query
          schema: { type: boolean }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertList"
  /alerts/{id}:
    get:
      operationId: getAlert
      tags: [alerts]
      summary: Get an alert by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAlert
      tags: [alerts]
      summary: Mark an alert as reviewed or escalated
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------
  # Audit logs
  # ---------------------------------------------------------------------
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    Role:
      type: string
      enum: [admin, compliance_officer, analyst, investigator]
    RiskLevel:
      type: string
      enum: [low, medium, high, critical]
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role:
          $ref: "#/components/schemas/Role"
        createdAt: { type: string, format: date-time }
    SignupInput:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string }
        password: { type: string, minLength: 6 }
        role:
          $ref: "#/components/schemas/Role"
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
...
    CentralityEntry:
      type: object
      required: [accountId, accountNo, customerName, degree, riskLevel]
      properties:
        accountId: { type: string }
        accountNo: { type: string }
        customerName: { type: string }
        degree: { type: integer }
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
...
    ShortestPathResult:
      type: object
      required: [found, path, length]
      properties:
        found: { type: boolean }
        path:
          type: array
          items: { type: string }
        length: { type: integer }
...
    Alert:
      type: object
      required:
        [
          id,
          customerId,
          customerName,
          accountNo,
          ruleTriggered,
          severity,
          description,
          createdAt,
          resolved,
        ]
      properties:
        id: { type: string }
        customerId: { type: string }
        customerName: { type: string }
        accountNo: { type: string }
        ruleTriggered: { type: string }
        severity:
          $ref: "#/components/schemas/RiskLevel"
        description: { type: string }
        createdAt: { type: string, format: date-time }
        resolved: { type: boolean }
    AlertUpdate:
      type: object
      properties:
        resolved: { type: boolean }
    AlertList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...